fix(scraper/fire-engine): decode responses using Content-Type/meta charset instead of forcing UTF-8
The fire-engine file-download path set status.content = buffer.toString("utf8"),
with a TODO acknowledging other encodings were unhandled. Non-UTF-8 pages
(Shift_JIS, windows-1251, ...) came back as mojibake.

The fetch engine already decodes correctly using the Content-Type header charset
with an HTML <meta charset> fallback. Extract that logic into a shared
decodeHtmlBuffer helper (verbatim, no behavior change for fetch) and use it in
fire-engine so both engines honor the declared charset and fall back to UTF-8.

Adds unit tests covering Shift_JIS via header charset, windows-1251 via meta
fallback, and the UTF-8 default.
