+++ apps/api/src/scraper/scrapeURL/engines/fetch/index.ts
@@ -8,80 +8,7 @@ import {
   InsecureConnectionError,
 } from "../utils/safeFetch";
 import { MockState, saveMock } from "../../lib/mock";
-import { TextDecoder } from "util";
-
-function decodeHtmlBuffer(
-  buf: Buffer,
-  contentType?: string,
-): {
-  text: string;
-  charset?: string;
-  charsetSource?: "header" | "meta";
-  decodeError?: unknown;
-} {
-  let text = buf.toString("utf8");
-
-  const headerCharsetRaw = (contentType?.match(
-    /charset\s*=\s*["']?([^;"'\s]+)/i,
-  ) ?? [])[1];
-  const headerCharset = headerCharsetRaw?.trim();
-
-  const metaCharsetRaw = (text.match(
-    /<meta\b[^>]*charset\s*=\s*["']?([^"'\s\/>]+)/i,
-  ) ?? [])[1];
-  const metaCharset = metaCharsetRaw?.trim();
-
-  if (headerCharset) {
-    try {
-      return {
-        text: new TextDecoder(headerCharset).decode(buf),
-        charset: headerCharset,
-        charsetSource: "header",
-      };
-    } catch (headerDecodeError) {
-      // If header charset is invalid/unsupported, fall back to meta charset.
-      if (
-        metaCharset &&
-        metaCharset.toLowerCase() !== headerCharset.toLowerCase()
-      ) {
-        try {
-          return {
-            text: new TextDecoder(metaCharset).decode(buf),
-            charset: metaCharset,
-            charsetSource: "meta",
-          };
-        } catch {
-          // Keep original header decode error for logging and utf8 fallback.
-        }
-      }
-      return {
-        text,
-        charset: headerCharset,
-        charsetSource: "header",
-        decodeError: headerDecodeError,
-      };
-    }
-  }
-
-  if (metaCharset) {
-    try {
-      return {
-        text: new TextDecoder(metaCharset).decode(buf),
-        charset: metaCharset,
-        charsetSource: "meta",
-      };
-    } catch (decodeError) {
-      return {
-        text,
-        charset: metaCharset,
-        charsetSource: "meta",
-        decodeError,
-      };
-    }
-  }
-
-  return { text };
-}
+import { decodeHtmlBuffer } from "../utils/decodeHtmlBuffer";
 
 export async function scrapeURLWithFetch(
   meta: Meta,
